requirements.txt
langchain_community
langchain
loguru
rich
sentence-transformers==2.3.0
unstructured
unstructured[pdf]
chromadb
rank_bm25
matplotlib
argparse
langchain-openai
pymupdf
streamlit
cohere
openai